remove dynamic routes
