+++ app/about/client.tsx
@@ -13,9 +13,9 @@ interface IVerify {
 }
 
 interface ISubscribe {
-  children: ReactNode
-  sitekey: string
-  verify: IVerify
+  children: ReactNode;
+  sitekey: string;
+  verify: IVerify;
 }
 
 const encode = (data: any) => {
@@ -31,8 +31,8 @@ interface IVerification {
 }
 
 export default function Subscribe({ children, sitekey, verify }: ISubscribe) {
-  const email = useRef<HTMLInputElement|null>(null);
-  const botField = useRef<HTMLInputElement|null>(null);
+  const email = useRef<HTMLInputElement | null>(null);
+  const botField = useRef<HTMLInputElement | null>(null);
   const ref = useRef<ReCAPTCHA>(null);
   const router = useRouter();
   // Uses same nomenclature as the Google API
@@ -83,39 +83,37 @@ export default function Subscribe({ children, sitekey, verify }: ISubscribe) {
   };
 
   return (
-        <form
-          className={style.form}
-          data-netlify="true"
-          data-netlify-honeypot="bot-field"
-          name={FORM_NAME}
-          onSubmit={onSubmit}
-        >
-          <p className={style.hidden}>
-            <label>{"this should be hidden, don't fill it out"}</label>
-            <input name="bot-field" ref={botField} />
-          </p>
-          <label htmlFor="email">
-            <code>Email</code>
-          </label>
-          <input
-            className={style.input}
-            id="email"
-            type="email"
-            name="email"
-            placeholder="..."
-            required
-            ref={email}
-          />
-          {
-            children
-          }
-           <button className={style.submit} type="submit" disabled={!verified}>
-              Subscribe
-            </button>
-          <Suspense fallback={<p>Loading ReCAPTCHA...</p>}>
-            {/* @ts-expect-error Server Component */}
-            <ReCAPTCHA ref={ref} sitekey={sitekey} onChange={setResponse} />
-          </Suspense>
-        </form>
+    <form
+      className={style.form}
+      data-netlify="true"
+      data-netlify-honeypot="bot-field"
+      name={FORM_NAME}
+      onSubmit={onSubmit}
+    >
+      <p className={style.hidden}>
+        <label>{"this should be hidden, don't fill it out"}</label>
+        <input name="bot-field" ref={botField} />
+      </p>
+      <label htmlFor="email">
+        <code>Email</code>
+      </label>
+      <input
+        className={style.input}
+        id="email"
+        type="email"
+        name="email"
+        placeholder="..."
+        required
+        ref={email}
+      />
+      {children}
+      <button className={style.submit} type="submit" disabled={!verified}>
+        Subscribe
+      </button>
+      <Suspense fallback={<p>Loading ReCAPTCHA...</p>}>
+        {/* @ts-expect-error Server Component */}
+        <ReCAPTCHA ref={ref} sitekey={sitekey} onChange={setResponse} />
+      </Suspense>
+    </form>
   );
 }
